feat: Add script to build and push container to ECR
This commit adds a new script, build_and_push_container.sh, which automates the process of building and pushing a Docker container to Amazon Elastic Container Registry (ECR). The script retrieves the AWS account ID, region, and algorithm name, and then creates the ECR repository if it doesn't exist. It also logs in to ECR, builds the Docker image locally, tags it with the full name, and pushes it to ECR.

Recent user commits:
- Update Dockerfile to comment mlio
- chore: Update requests library to version 2.32.3
- Update Dockerfile to use Miniforge3 for Conda installation
- Update Dockerfile to use Miniforge3 for Conda installation
- feat: Update Dockerfile to use Miniforge3 for Conda installation
- Update Dockerfile to use Miniforge3 for Conda installation
- update Miniconda version to 24.5.0-0 and Conda checksum
- updating dockerfile
- Upgrading to Python 3.10 and XGBoost 2.1-1
- bump python version (#432)